Refactor: Change price to BigDecimal in tables, remove migration V11

+++ aurora-backend/src/main/resources/db/migration/V3__create_book_related_tables.sql
@@ -2,7 +2,7 @@ CREATE TABLE book(
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(255) NOT NULL,
     description VARCHAR(500),
-    price FLOAT,
+    price DECIMAL(10,2),
     image_url VARCHAR(255),
     stock INT NOT NULL DEFAULT 0
 );
